Big clean-up part 1.

+++ src/entity/Bullet.scala
@@ -4,32 +4,31 @@ import org.newdawn.slick.{Image, Color}
 
 class Bullet(var x: Float, var y: Float, color: Color, speedX: Float = 0f, speedY: Float = -1.3f, angle: Float = 0f) extends Entity {
   private val image = new Image("gfx/bullet.png")
+  private val marginX = image.getWidth / 2f
+  private val marginY = image.getHeight / 2f
   image.setRotation(angle)
-  private val imageAlignX = image.getWidth / 2f
-  private val imageAlignY = image.getHeight / 2f
 
   def update(delta: Int, linker: Linker) {
-    y += speedY * delta
-    x += speedX * delta
+    for (i <- 0 until delta) {
+      y += speedY
+      x += speedX
+    }
 
-    if (linker.getReference.isEmpty) sys.error("Could not get linker reference from Bullet.")
-    else {
-      val collisionX = (x - imageAlignX) -> (x + imageAlignX)
-      val collisionY = (y - imageAlignY + 20) -> (y + imageAlignY + 20)
+    if (linker.reference(0).isDefined) {
+      val collisionX = (x - marginX) -> (x + marginX)
+      val collisionY = (y - marginY + 20) -> (y + marginY + 20)
 
-      var currentItem = linker.getReference.get.getNext
+      var currentItem = linker.reference(0).get.getNext
       while (currentItem.isDefined) {
-        if (currentItem.get.collision(collisionX, collisionY))
-          unlink()
+        if (currentItem.get.collision(collisionX, collisionY)) unlink()
         currentItem = currentItem.get.getNext
       }
     }
 
-    if (y < -100)
-      unlink()
+    if (y < -100 || y > 1000 || x < -100 || x > 1500) unlink()
   }
 
   def render() {
-    image.draw(x - imageAlignX, y - imageAlignY, color)
+    image.draw(x - marginX, y - marginY, color)
   }
 }
